requirements.txt
aiogram~=3.6.0
aiogram_calendar==0.5.0
apscheduler==3.9.1
sqlalchemy==1.4.41
python-dotenv==0.19.2
python-dateutil==2.8.2